+++ cognite/neat/rules/models/dms/_rules_input.py
@@ -11,8 +11,10 @@ from cognite.neat.rules.models.entities import (
     ClassEntity,
     ContainerEntity,
     DMSUnknownEntity,
+    EdgeEntity,
+    ReverseConnectionEntity,
     ViewEntity,
-    ViewPropertyEntity,
+    load_connection,
     load_dms_value_type,
 )
 
@@ -79,12 +81,12 @@ class DMSInputMetadata(InputComponent[DMSMetadata]):
 class DMSInputProperty(InputComponent[DMSProperty]):
     view: str
     view_property: str | None
-    value_type: str | DataType | ViewPropertyEntity | ViewEntity | DMSUnknownEntity
-    property_: str | None
+    value_type: str | DataType | ViewEntity | DMSUnknownEntity
+    property_: str | None = None
     class_: str | None = None
     name: str | None = None
     description: str | None = None
-    connection: Literal["direct", "edge", "reverse"] | None = None
+    connection: Literal["direct"] | ReverseConnectionEntity | EdgeEntity | str | None = None
     nullable: bool | None = None
     immutable: bool | None = None
     is_list: bool | None = None
@@ -103,6 +105,7 @@ class DMSInputProperty(InputComponent[DMSProperty]):
         output = super().dump()
         output["View"] = ViewEntity.load(self.view, space=default_space, version=default_version)
         output["Value Type"] = load_dms_value_type(self.value_type, default_space, default_version)
+        output["Connection"] = load_connection(self.connection, default_space, default_version)
         output["Property (linage)"] = self.property_ or self.view_property
         output["Class (linage)"] = (
             ClassEntity.load(self.class_ or self.view, prefix=default_space, version=default_version)
@@ -125,6 +128,7 @@ class DMSInputContainer(InputComponent[DMSContainer]):
     description: str | None = None
     reference: str | None = None
     constraint: str | None = None
+    used_for: Literal["node", "edge", "all"] | None = None
 
     @classmethod
     def _get_verified_cls(cls) -> type[DMSContainer]:
@@ -158,6 +162,7 @@ class DMSInputContainer(InputComponent[DMSContainer]):
             name=container.name or None,
             description=container.description,
             constraint=", ".join(constraints) or None,
+            used_for=container.used_for,
         )